python qsm

+++ src/main/kotlin/com/qsm/ad/repositroys/BaseRepository.java
@@ -8,6 +8,6 @@ import org.springframework.data.repository.NoRepositoryBean;
  * Created by think on 2017/7/30.
  */
 @NoRepositoryBean
-public interface BaseRepository<T> extends JpaRepository<T, Long>, JpaSpecificationExecutor<T> {
+public interface BaseRepository<T> extends JpaRepository<T, Integer>, JpaSpecificationExecutor<T> {
 
 }
